requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.8.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.1.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
dnspython==2.7.0
    # via email-validator
email-validator==2.2.0
    # via fastapi
exceptiongroup==1.2.2
    # via anyio
fastapi[standard]==0.115.8
    # via -r requirements.in
fastapi-cli[standard]==0.0.7
    # via fastapi
flatbuffers==25.2.10
    # via onnxruntime
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via fastapi
humanfriendly==10.0
    # via coloredlogs
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
imageio==2.37.0
    # via scikit-image
jinja2==3.1.5
    # via fastapi
jsonschema==4.23.0
    # via rembg
jsonschema-specifications==2024.10.1
    # via jsonschema
lazy-loader==0.4
    # via scikit-image
llvmlite==0.44.0
    # via numba
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via scikit-image
numba==0.61.0
    # via pymatting
numpy==2.1.3
    # via
    #   imageio
    #   numba
    #   onnxruntime
    #   opencv-python-headless
    #   pymatting
    #   rembg
    #   scikit-image
    #   scipy
    #   tifffile
onnxruntime==1.20.1
    # via -r requirements.in
opencv-python-headless==4.11.0.86
    # via rembg
packaging==24.2
    # via
    #   lazy-loader
    #   onnxruntime
    #   pooch
    #   scikit-image
pillow==11.1.0
    # via
    #   -r requirements.in
    #   imageio
    #   pymatting
    #   rembg
    #   scikit-image
platformdirs==4.3.6
    # via pooch
pooch==1.8.2
    # via rembg
protobuf==5.29.3
    # via onnxruntime
pydantic==2.10.6
    # via fastapi
pydantic-core==2.27.2
    # via pydantic
pygments==2.19.1
    # via rich
pymatting==1.1.13
    # via rembg
pyside6==6.8.2.1
    # via -r requirements.in
pyside6-addons==6.8.2.1
    # via pyside6
pyside6-essentials==6.8.2.1
    # via
    #   pyside6
    #   pyside6-addons
python-dotenv==1.0.1
    # via uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rembg==2.0.63
    # via -r requirements.in
requests==2.32.3
    # via pooch
rich==13.9.4
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.13.2
    # via fastapi-cli
rpds-py==0.23.1
    # via
    #   jsonschema
    #   referencing
scikit-image==0.25.2
    # via rembg
scipy==1.15.2
    # via
    #   pymatting
    #   rembg
    #   scikit-image
shellingham==1.5.4
    # via typer
shiboken6==6.8.2.1
    # via
    #   pyside6
    #   pyside6-addons
    #   pyside6-essentials
sniffio==1.3.1
    # via anyio
starlette==0.45.3
    # via fastapi
sympy==1.13.3
    # via onnxruntime
tifffile==2025.2.18
    # via scikit-image
tqdm==4.67.1
    # via rembg
typer==0.15.1
    # via fastapi-cli
typing-extensions==4.12.2
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   referencing
    #   rich
    #   rich-toolkit
    #   typer
    #   uvicorn
urllib3==2.3.0
    # via requests
uvicorn[standard]==0.34.0
    # via
    #   fastapi
    #   fastapi-cli
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.4
    # via uvicorn
websockets==15.0
    # via uvicorn
